+++ next.config.ts
@@ -1,15 +1,7 @@
 import type { NextConfig } from 'next';
 
 const nextConfig: NextConfig = {
-  async redirects() {
-    return [
-      {
-        source: '/',
-        destination: '/industry/products',
-        permanent: true, // 永続的なリダイレクト（301）
-      },
-    ];
-  },
+  /* config options here */
 };
 
 export default nextConfig;
