+++ src/app/industry/products/[id]/page.tsx
@@ -1,0 +1,8 @@
+export default function Page() {
+  return (
+    <>
+      <h2>商品在庫管理</h2>
+      <p>商品在庫の一覧を表示する</p>
+    </>
+  );
+}
