+++ src/app/layout.tsx
@@ -1,20 +1,15 @@
 import type { Metadata } from 'next';
-import { Geist, Geist_Mono } from 'next/font/google';
-import './globals.css';
+import { M_PLUS_Rounded_1c } from 'next/font/google';
 
-const geistSans = Geist({
-  variable: '--font-geist-sans',
+const mPlusRounded1c = M_PLUS_Rounded_1c({
   subsets: ['latin'],
-});
-
-const geistMono = Geist_Mono({
-  variable: '--font-geist-mono',
-  subsets: ['latin'],
+  weight: ['400', '500', '700'],
+  display: 'swap',
 });
 
 export const metadata: Metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
+  title: '在庫管理システム',
+  description: '実装で学ぶフルスタックWeb開発で作る「在庫管理システム」',
 };
 
 export default function RootLayout({
@@ -23,12 +18,8 @@ export default function RootLayout({
   children: React.ReactNode;
 }>) {
   return (
-    <html lang="en">
-      <body
-        className={`${geistSans.variable} ${geistMono.variable} antialiased`}
-      >
-        {children}
-      </body>
+    <html lang="ja" className={`${mPlusRounded1c.className}`}>
+      <body>{children}</body>
     </html>
   );
 }
