+++ src/app/login/page.tsx
@@ -1,0 +1,8 @@
+export default function Page() {
+  return (
+    <>
+      <h2>ログイン</h2>
+      <p>ログインの入力項目を表示する</p>
+    </>
+  );
+}
